Refactor out last hook-style GQL query and fix imag name
